+++ src/main/java/com/ll/commars/domain/community/board/repository/HashTagRepository.java
@@ -1,0 +1,9 @@
+package com.ll.commars.domain.community.board.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.ll.commars.domain.community.board.entity.HashTag;
+
+public interface HashTagRepository extends JpaRepository<HashTag, Long> {
+
+}
