--- src/main/java/com/ll/commars/domain/community/boardHashTag/repository/HashTagRepository.java
@@ -1,21 +1,0 @@
-package com.ll.commars.domain.community.boardHashTag.repository;
-/*
-import com.ll.commars.domain.community.board.entity.Board;
-import com.ll.commars.domain.community.boardHashTag.entity.HashTag;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-
-public interface HashTagRepository extends JpaRepository<HashTag, Long> {
-    // 특정 태그로 검색
-    List<HashTag> findByTag(String tag);
-
-    // 게시글 관련 태그 조회
-    List<HashTag> findByBoard(Board board);  // Board와 연결된 HashTag 조회
-
-    // 게시글 관련 태그 삭제
-    void deleteByBoard(Board board);
-
-
-}
-*/
